Remove ApiParamsBuilder
